+++ flutter/lib/src/integrations/on_error_integration.dart
@@ -3,6 +3,7 @@ import 'dart:ui';
 
 import 'package:flutter/widgets.dart';
 import 'package:sentry/sentry.dart';
+import '../binding_utils.dart';
 import '../sentry_flutter_options.dart';
 
 typedef ErrorCallback = bool Function(Object exception, StackTrace stackTrace);
@@ -28,9 +29,15 @@ class OnErrorIntegration implements Integration<SentryFlutterOptions> {
   @override
   void call(Hub hub, SentryFlutterOptions options) {
     _options = options;
+    final binding = BindingUtils.getWidgetsBindingInstance();
+
+    if (binding == null) {
+      return;
+    }
+
+    // WidgetsBinding works with WidgetsFlutterBinding and other custom bindings
     final wrapper = dispatchWrapper ??
-        // WidgetsBinding works with WidgetsFlutterBinding and other custom bindings
-        PlatformDispatcherWrapper(WidgetsBinding.instance.platformDispatcher);
+        PlatformDispatcherWrapper(binding.platformDispatcher);
 
     if (!wrapper.isOnErrorSupported(options)) {
       return;